requirements.txt
# Core data processing
polars>=0.20.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0  # For reading Excel files
xlsxwriter>=3.2.9

# NLP and ML
transformers>=4.35.0
torch>=2.0.0
sentence-transformers>=2.2.0
spacy>=3.7.0
accelerate>=0.24.0
bitsandbytes>=0.41.0  # For quantization

# Topic modeling (from reference script)
bertopic>=0.16.0
nltk>=3.8.0
scikit-learn>=1.3.0
# TurfTopic for multilingual topic modeling (if available)
# turftopic>=0.1.0  # Uncomment when TurfTopic is available on PyPI

# Database
psycopg2-binary>=2.9.9  # PostgreSQL adapter
sqlalchemy>=2.0.0
asyncpg>=0.29.0  # Async PostgreSQL

# Web framework & API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Frontend/Dashboard
streamlit>=1.28.0
plotly>=5.17.0
altair>=5.1.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pyyaml>=6.0.1
paramiko>=3.0.0  # For SSH remote processing
